trx_toolkit/*: Don't use `del x; x = None` idiom
In Python it is enough to do `x = None` to release the object that was
pointed to by x previously. So at Python level doing this extra
del only spends more CPU time, but when we will switch to Cython, and
e.g. Msg.burst will be a C-level attribute, the following will not work
at all

	del msg.burst
	msg.burst = None

because at runtime it will complain that

        del msg.burst
    AttributeError: 'Msg' object has no attribute 'burst' and no __dict__ for setting new attributes

-> Remove unneeded del to save some time and avoid problems with upcoming switch to Cython.

Change-Id: I7a83bdd52fb9318bd8b975f85ce37c7144873f61