trx_toolkit/udp_link: Factor code to describe remote into .desc_remote() function
And use that utility everywhere where remote of UDPLink is logged.

The reason we are doing this is that with upcoming switch to Cython the
way remote address is stored will change to `struct sockaddr_in` and
instead of updating all users, we will need to only change
UDPLink.desc_remote() in one place.

Add .desc_local() for symmetry.

Change-Id: I1e2fa560ada7a8de4c9b9150058c2a1c73874fbe